+++ raster_tools/extract.py
@@ -527,6 +527,25 @@ class ThreeDiAHN3HHNK(Operation, ThreeDiBase):
         super(ThreeDiAHN3HHNK, self).__init__(floor, **kwargs)
 
 
+class ThreeDiAHN3Almere(Operation, ThreeDiBase):
+    name = '3di-ahn3-almere'
+
+    def __init__(self, floor, **kwargs):
+        """ Initialize the operation. """
+        # first define the layers
+        self.layers = {
+            self.I_BATHYMETRY: dict(layers=','.join([
+                'intern:nl:ahn3:int-almere',
+                'intern:nl:ahn3:bag-almere!{}'.format(floor),
+                'dem:water',
+            ])),
+            self.I_LANDUSE: dict(layers='cover:3di'),
+            self.I_SOIL: dict(layers='soil:3di'),
+        }
+        # then let the base class handle the rest
+        super(ThreeDiAHN3Almere, self).__init__(floor, **kwargs)
+
+
 class ThreeDiRD(Operation, ThreeDiBase):
     name = '3di-rd'