+++ CHANGES.rst
@@ -5,17 +5,19 @@ Changelog of raster-tools
 0.2 (unreleased)
 ----------------
 
-- Extended extract tool with 3Di-AHN3 option.
+- Extended extract tool with 3di-ahn3 option.
 
 - Added meta about raster-server layers to extract tool.
 
 - Leanified and documented AHN3 downloader.
 
-- Extended extract tool with 3Di-AHN3-HHNK option.
+- Extended extract tool with 3di-ahn3-hhnk option.
 
-- Extended extract tool with 3Di-RD option.
+- Extended extract tool with 3di-rd option.
 
 - Add difference script from the past.
+
+- Extended extract tool with 3Di-ahn3-almere option.
 
 
 0.1 (2016-12-12)